Rename Persistence/ to Adapters/
The one class in there is not really a persistence class.
Let's lump all non-presentation Adapters together for now.
